[fix] 리팩토링 진행 중
